requirements.txt
anyio==3.7.0
bcrypt==4.0.1
black==23.3.0
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
docopt==0.6.2
email-validator==2.0.0.post2
fastapi==0.97.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.6.0
pluggy==1.0.0
psycopg2==2.9.6
pydantic==1.10.9
PyJWT==2.7.0
pytest==7.3.2
pytest-watch==4.2.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
watchdog==3.0.0
